backend/requirements.txt
# FastAPI & ASGI Server
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0

# MongoDB Async Driver
motor>=3.6.0,<3.7.0

# Data Validation & Settings
pydantic[email]>=2.10.0,<2.11.0
pydantic-settings>=2.7.0,<2.8.0

# Authentication
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.0,<1.8.0
python-multipart>=0.0.18

# Machine Learning
scikit-learn>=1.6.0,<1.7.0
numpy>=2.1.0,<2.3.0
pandas>=2.2.0,<2.3.0

# WebSockets
websockets>=14.0,<15.0

# Utilities
python-dateutil>=2.9.0

google-generativeai
